% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_all_listes}
\alias{get_all_listes}
\title{~ api : recuperer toutes les listes sur une thematique}
\usage{
get_all_listes(theme, def_url = path.package("nomensland"))
}
\arguments{
\item{theme}{Character. Nom de la thématique recherchée
(par exemple : \code{"Chirurgie bariatrique"}).}

\item{def_url}{Character. Chemin vers les données du package.
Par défaut, utilise \code{path.package("nomensland")}.
Ce paramètre est principalement destiné à un usage interne
ou pour des environnements spécifiques.}
}
\description{
~ api : recuperer toutes les listes sur une thematique
}
\examples{
\dontrun{
get_all_listes("Chirurgie bariatrique")
get_all_listes("Recours Exceptionnel")
}

}
