% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_dictionnaire_listes}
\alias{get_dictionnaire_listes}
\title{~ api : recuperer le dictionnaire des listes}
\usage{
get_dictionnaire_listes(def_url = path.package("nomensland"))
}
\arguments{
\item{def_url}{Character. Chemin vers les données du package.
Par défaut, utilise \code{path.package("nomensland")}.
Ce paramètre est principalement destiné à un usage interne
ou pour des environnements spécifiques.}
}
\description{
~ api : recuperer le dictionnaire des listes
}
\examples{
\dontrun{
get_dictionnaire_listes()
}


}
\author{
G. Pressiat
}
