% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_dictionnaire_tables}
\alias{get_dictionnaire_tables}
\title{~ api : avoir le dictionnaire des referentiels}
\usage{
get_dictionnaire_tables(def_url = path.package("nomensland"))
}
\arguments{
\item{def_url}{Character. Chemin vers les données du package.
Par défaut, utilise \code{path.package("nomensland")}.
Ce paramètre est principalement destiné à un usage interne
ou pour des environnements spécifiques.}
}
\description{
~ api : avoir le dictionnaire des referentiels
}
\examples{
\dontrun{
get_dictionnaire_tables()
}


}
\author{
G. Pressiat
}
