% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_table}
\alias{get_table}
\title{~ api : recuperer une table (referentiel)}
\usage{
get_table(table, version = "", def_url = path.package("nomensland"))
}
\arguments{
\item{table}{Character. Nom de la table.}

\item{version}{Character. Version pour filtrer la table (ex: 2025).}

\item{def_url}{Character. Chemin vers les données du package.
Par défaut, utilise \code{path.package("nomensland")}.
Ce paramètre est principalement destiné à un usage interne
ou pour des environnements spécifiques.}
}
\description{
~ api : recuperer une table (referentiel)
}
\examples{
\dontrun{
get_table('dictionnaire_tables')
get_table('ccam_actes')
get_table('cim', 2016:2017)
get_table('tarifs_mco_ghs', 2015:2018)
}

}
\author{
G. Pressiat
}
