% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{explore_nomensland}
\alias{explore_nomensland}
\title{Explorer les tables avec shiny}
\usage{
explore_nomensland(
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\description{
Explorer les tables avec shiny
}
