% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_dictionnaire_tables}
\alias{get_dictionnaire_tables}
\title{~ api : avoir le dictionnaire des referentiels}
\usage{
get_dictionnaire_tables(def_url = path.package("nomensland"))
}
\description{
~ api : avoir le dictionnaire des referentiels
}
\examples{
\dontrun{
get_dictionnaire_tables()
}

}
\author{
G. Pressiat
}
