% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_liste}
\alias{get_liste}
\title{~ api : recuperer une liste}
\usage{
get_liste(nom_liste, def_url = path.package("nomensland"))
}
\description{
~ api : recuperer une liste
}
\examples{
\dontrun{
get_liste('chip')
}

}
\author{
G. Pressiat
}
