% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_table}
\alias{get_table}
\title{~ api : recuperer une table (referentiel)}
\usage{
get_table(table, version = "", def_url = path.package("nomensland"))
}
\description{
~ api : recuperer une table (referentiel)
}
\examples{
\dontrun{
get_table('dictionnaire_tables')
get_table('ccam_actes')
get_table('cim', 2016:2017)
get_table('tarifs_mco_ghs', 2015:2018)
}

}
\author{
G. Pressiat
}
